V3.0.0 (#40)
* Activity category permission change

Added everyone and guest role deny view permissions to all new activitiy categories.

* Channel topics and names with emojis

Added channel topics to all created text channels by the bot. Also added emojis to all channel names created by the bot. Special treat to coffe chat and among us.

* Stamp time increase

Increased the time poeple have to stamp to one minute, before it was at 20 seconds.

* Pin important messages

Pinned important messages with emoji collectors.

* archive command

Added a new archive activity command, it will remove all channels except for the text channel which gets moved to the archive category. This category is created if not found.

* Revert "NC-archive"

* New command -> archive

Will remove all activity channels except for the text channel. It is moved to the archive category. This category is created if not found.

* Created archive command and added to new activity

Archive command will remove all channels except for text, gets moved to archive category. This category is created if not found!

* Removed uncessesary await

discordServices.checkForRole is not async anymore, so removed all unecessary awaits.

* Fixes new naming conventions

Moved the voice and text channel naming convention to discordServices, for it to be easier to edit.

* U workshop ta (#31)

* Automatic message deletion

Bot now automatically deletes messages sent to a list of channels in discordServices after a few seconds. This is then used to keep channels like workshop assistance clear.

* Bug fix and QOL improvements

Small bug fixes and quality of life improvements on the ta functionality.

* More bug fixes

Some bug fixes on archive, newactivity and discordServices.

* More bug fixes and namig change

Changed the name of new voice channels to not use activity name

* Voice channel changes

Made voice channels be:
- public when activity is activity
- private when activity is workshop, among us or coffee chat
- the add voice channel emoji will follow these rules

* Bug fixes and among us improvements

Bug fix for coffee chats, among us emoji reaction will now make all previouse and new voice channels except general channel have a limit of 12 users.

* Mentor upgrade

Add the mentor request a ticket channel to the black list.

* Added TA low tech solution

Added an emoji collector to the information message in the TA channel. This is a one way command, no coming back! Hides the general voice channel and disables the pull in functionality.

* BF-startm

startm was not grabbing already made categories due to the emoji addition. We now look for a name that ends with the special string, no emojis used. Also added topics to channesl created by this command.

* U help and clearchat

They both use the same procedure to get commands to send. clearchat will send the hacker commands for any regular channel.

* U clearchat not remove pinned messages

clearchat has a new boolean argument, true will cause pinned messages to not be deleted, false will just remove all messages.

* Consolidate commands in one group

Moved all start... commands to one new group called a_start_commands. The sponsor boothing command is not there right now, waitng for revamp.

* U new member message

Imrpoved the new member DM by using an embed with fields.

* U help

The help command can now be used out of the guild!

* U stamp collection time

stamp collection and text channels that explain it use a discordServices variable. Added a command to change this value, text apperances will not update unless they are new.

* minor text changes

* QOL and Better mentor support (#34)

Quality of Life improvements to startm and added more mentor per ticket functionality. Once a ticket is taken by a mentor other mentors can join!

* Nc hide unhide (#35)

* NC hide/unhide, added to console, helper function

Added the new hide/unhide command, it has some limitations, can only be used once or twice, depending how the activity starts. Added the command to the console. Changed a helper function to make it more usefull.

* BF remove activity

quick bug fix by adding an await.

* QOL and Better mentor support (#36)

Quality of Life improvements to startm and added more mentor per ticket functionality. Once a ticket is taken by a mentor other mentors can join!

* Bf 3 (#38)

* QOL verify and attend

Use of embeds for verify and attend. Also improved text.

* QOL startm

Text improvements for ticket system.

* U and QOL Permissions

Permissions now work! We use update instead of set permissions so that they hold the category's permissions!

* BF Ticket Permissions

Fixed some bugs with ticket permissions. It all works now!

* U Keep Archive Bottom

We now keep the archive on the very bottom!

* BF startm

Fixed a bug realted to undeclared variable due to previouse PR.

* BF Permission REMOVED HIDE/UNHIDE

The hide /unhide functionality is bringing a lot of issues with permissions and  category locking and unlocking so I have decided that the best option is just to remove it. The code is still present but it is disabled and shouldnt be used!

* BF Passowrd Stamp and Workshop QOL

Bug fixes for password stamp, changes TA assistance one-liner timer to 20 seconds. Embed improvements for polls.

* Bf 3 (#39)

* QOL verify and attend

Use of embeds for verify and attend. Also improved text.

* QOL startm

Text improvements for ticket system.

* U and QOL Permissions

Permissions now work! We use update instead of set permissions so that they hold the category's permissions!

* BF Ticket Permissions

Fixed some bugs with ticket permissions. It all works now!

* U Keep Archive Bottom

We now keep the archive on the very bottom!

* BF startm

Fixed a bug realted to undeclared variable due to previouse PR.

* BF Permission REMOVED HIDE/UNHIDE

The hide /unhide functionality is bringing a lot of issues with permissions and  category locking and unlocking so I have decided that the best option is just to remove it. The code is still present but it is disabled and shouldnt be used!

* BF Passowrd Stamp and Workshop QOL

Bug fixes for password stamp, changes TA assistance one-liner timer to 20 seconds. Embed improvements for polls.

* U ask and QOL team foramtion

A bit better otpimized ask command and text improvements to team formation!

* changed emojis on Ta console, added stamp roles

* QOL Text and BF create channels

Text improvements all around and BF to create channel.

* text change to password-stamp

* BF startm and polls, NC showi

Bug fixes on startm and activity polls. New command to send infomrative message about verify or attend.

+++ discord-services.js
@@ -23,6 +23,15 @@ var stamp9Role = '777163468053938186';
 var stamp10Role = '777163488019480586';
 var stamp11Role = '777163506902237196';
 var stamp12Role = '777163524568776704';
+var stamp12Role = '777163524568776704';
+var stamp13Role = '784224112909221948';
+var stamp14Role = '784224898230779945';
+var stamp15Role = '784224924633923635';
+var stamp16Role = '784224924633923635';
+var stamp17Role = '784224964001005589';
+var stamp18Role = '784224981386133525';
+var stamp19Role = '784224999698726942';
+var stamp20Role = '784225017172590622';
 module.exports.everyoneRole = everyoneRole;
 module.exports.hackerRole = hackerRole;
 module.exports.guestRole = guestRole;
@@ -44,6 +53,15 @@ module.exports.stamp9Role = stamp9Role;
 module.exports.stamp10Role = stamp10Role;
 module.exports.stamp11Role = stamp11Role;
 module.exports.stamp12Role = stamp12Role;
+module.exports.stamp13Role = stamp13Role;
+module.exports.stamp14Role = stamp14Role;
+module.exports.stamp15Role = stamp15Role;
+module.exports.stamp16Role = stamp16Role;
+module.exports.stamp17Role = stamp17Role;
+module.exports.stamp18Role = stamp18Role;
+module.exports.stamp19Role = stamp19Role;
+module.exports.stamp20Role = stamp20Role;
+
 
 // other project wide vars
 var embedColor = '#26fff4';
@@ -56,6 +74,14 @@ var tfTeamEmbedColor = '#1929ff';
 module.exports.tfTeamEmbedColor = tfTeamEmbedColor;
 var tfHackerEmbedColor = '#ff33f1';
 module.exports.tfHackerEmbedColor = tfHackerEmbedColor;
+var specialDMEmbedColor = '#fc6b03';
+module.exports.specialDMEmbedColor = specialDMEmbedColor;
+
+const blackList = new Map();
+module.exports.blackList = blackList;
+
+var stampCollectTime = 60;
+module.exports.stampCollectTime = stampCollectTime;
 
 // Common channels
 
@@ -115,6 +141,17 @@ var incomingReportChannel = '780305617267982366';
 module.exports.incomingReportChannel = incomingReportChannel;
 
 
+// naming conventions
+
+var activityTextChannelName = 'activity-banter';
+module.exports.activityTextChannelName = activityTextChannelName;
+
+var activityVoiceChannelName = 'activity-room';
+module.exports.activityVoiceChannelName = activityVoiceChannelName;
+
+
+// helper function
+
 // Checks if the memeber has a role, returns true if it does
 function checkForRole(member, role) {
     return member.roles.cache.has(role);
@@ -170,7 +207,7 @@ function isAdminConsole(channel) {
 module.exports.isAdminConsole = isAdminConsole;
 
 // will add given number of voice channels to the given activity, the category object of the activity is necessary
-async function addVoiceChannelsToActivity(activityName, number, category, channelManager, maxUsers = 0) {
+async function addVoiceChannelsToActivity(activityName, number, category, channelManager, isPrivate, maxUsers = 0) {
     // udpate db and get total number of channels
     var total = await firebaseActivity.addVoiceChannels(activityName, number);
 
@@ -179,30 +216,15 @@ async function addVoiceChannelsToActivity(activityName, number, category, channe
 
     // create voice channels
     for (; index < total; index++) {
-        channelManager.create(activityName + '-' + index, {type: 'voice', parent: category, permissionOverwrites : [
-            {
-                id: hackerRole,
-                deny: ['VIEW_CHANNEL'],
-            },
-            {
-                id: attendeeRole,
-                deny: ['VIEW_CHANNEL'],
-                allow: ['USE_VAD', 'SPEAK'],
-            },
-            {
-                id: sponsorRole,
-                deny: ['VIEW_CHANNEL'],
-            },
-            {
-                id: mentorRole,
-                allow: ['VIEW_CHANNEL', 'USE_VAD', 'SPEAK', 'MOVE_MEMBERS'],
-            },
-            {
-                id: staffRole,
-                allow: ['VIEW_CHANNEL', 'USE_VAD', 'SPEAK', 'MOVE_MEMBERS'],
-            }
-        ],
-        userLimit: maxUsers === 0 ? undefined : maxUsers}).catch(console.error);
+        channelManager.create('🔊Room' + '-' + index, {
+            type: 'voice', 
+            parent: category, 
+            userLimit: maxUsers === 0 ? undefined : maxUsers
+        }).then(channel => {
+            channel.updateOverwrite(attendeeRole, {VIEW_CHANNEL: isPrivate ? false : true, USE_VAD: true, SPEAK: true});
+            channel.updateOverwrite(sponsorRole, {VIEW_CHANNEL: isPrivate ? false : true, USE_VAD: true, SPEAK: true});
+            channel.updateOverwrite(mentorRole, {MOVE_MEMBERS: true, USE_VAD: true});
+        }).catch(console.error);
     }
 
     return total;
@@ -225,19 +247,58 @@ async function removeVoiceChannelsToActivity(activityName, number, category){
     // we remove one because we are counting from 0
     // remove voice channels
     for (var index = total - 1; index >= final; index--) {
-        var channelName = activityName + '-' + index;
-        var channel = await category.children.find(channel => channel.name === channelName);
-        channel.delete();
+        var channelName = '🔊Room' + '-' + index;
+        var channel = await category.children.find(channel => channel.name.endsWith(channelName));
+        if (channel != undefined) {
+            channel.delete();
+        }
     }
 
     return final;
 }
 module.exports.removeVoiceChannelsToActivity = removeVoiceChannelsToActivity;
 
+// will make all voice channels except the general one private to attendees and sponsors
+async function changeVoiceChannelPermissions(activityName, category, toHide) {
+    // udpate db and get total number of channels
+    var total = await firebaseActivity.numOfVoiceChannels(activityName);
+
+    // grab index where channel naming should stampt, in case there are already channels made
+    // we remove one because we are counting from 0
+    // remove voice channels
+    for (var index = total - 1; index >= 0; index--) {
+        var channelName = 'Room' + '-' + index;
+        var channel = await category.children.find(channel => channel.name.endsWith(channelName));
+        if (channel != undefined) {
+            channel.updateOverwrite(attendeeRole, {VIEW_CHANNEL: toHide ? false : true});
+            channel.updateOverwrite(sponsorRole, {VIEW_CHANNEL: toHide ? false : true});
+        }
+    }
+}
+module.exports.changeVoiceChannelPermissions = changeVoiceChannelPermissions;
+
+// will add a max amount of users to the activity voice channels
+async function addLimitToVoiceChannels(activityName, category, limit) {
+    // udpate db and get total number of channels
+    var total = await firebaseActivity.numOfVoiceChannels(activityName);
+
+    // grab index where channel naming should stampt, in case there are already channels made
+    // we remove one because we are counting from 0
+    // remove voice channels
+    for (var index = total - 1; index >= 0; index--) {
+        var channelName = '🔊Room' + '-' + index;
+        var channel = await category.children.find(channel => channel.name.endsWith(channelName));
+        if (channel != undefined) {
+            await channel.edit({userLimit: limit});
+        }
+    }
+}
+module.exports.addLimitToVoiceChannels = addLimitToVoiceChannels;
+
 // deletes a message if the message hasn't been deleted already
-function deleteMessage(message) {
+function deleteMessage(message, timeout = 0) {
     if (message.deleted === false) {
-        message.delete();
+        message.delete({timeout: timeout});
     }
 }
 module.exports.deleteMessage = deleteMessage;
